+++ Runtime/Interactables/InteractableHandler.cs
@@ -4,44 +4,32 @@ using UnityEngine.Events;
 namespace OneM.InteractableSystem
 {
     /// <summary>
-    /// Low coupling implementation of <see cref="IInteractable"/> using <see cref="UnityEvent"/>.
+    /// Low coupling implementation of <see cref="IInteractable"/> using serialized Unity Events.
     /// </summary>
     /// <remarks>
     /// Use this component to quickly create an Interactable instance reacting when 
     /// <see cref="AbstractInteractor{T}"/> implementations detect collisions.
     /// </remarks>
     [DisallowMultipleComponent]
-    public sealed class InteractableUnityEvent : MonoBehaviour, IInteractable
+    public sealed class InteractableHandler : MonoBehaviour, IInteractable
     {
         [field: SerializeField]
         public Collider Collider { get; private set; }
 
-        /// <summary>
-        /// Event fired when interacted with this object.
-        /// </summary>
         [Header("EVENTS")]
-        public UnityEvent<Transform> OnInteracted;
-
-        /// <summary>
-        /// Event fired when this object availability is changed.
-        /// </summary>
-        [Space]
+        [Tooltip("Event fired when this object availability is changed. Use it to show the Interaction Input.")]
         public UnityEvent<bool> OnAvailabilityChanged;
-
-        /// <summary>
-        /// Event fired when interacted with this object fails.
-        /// </summary>
-        [Space]
+        [Tooltip("Event fired when interacted with this object and it success.")]
+        public UnityEvent<Transform> OnInteracted;
+        [Tooltip("Event fired when interacted with this object and it fails.")]
         public UnityEvent OnInteractionFailed;
-
-        private void Reset() => Collider = GetComponent<Collider>();
 
         public bool CanInteract() => enabled;
         public bool CanCollide() => CanInteract();
         public void Interact(Transform interactor) => OnInteracted?.Invoke(interactor);
         public void ShowInteractionFail() => OnInteractionFailed?.Invoke();
         public void ChangeAvailability(bool isAvailable) => OnAvailabilityChanged?.Invoke(isAvailable);
-        public void EnterCollision(Transform interactor) => ChangeAvailability(true);
-        public void ExitCollision(Transform interactor) => ChangeAvailability(false);
+        public void EnterCollision(Transform _) => ChangeAvailability(true);
+        public void ExitCollision(Transform _) => ChangeAvailability(false);
     }
 }